more cleaning before merge
